wip: sableye encounter
