--- app/public/src/game/components/unown-manager.ts
@@ -1,206 +1,0 @@
-import { GameObjects } from "phaser"
-import { Transfer } from "../../../../types"
-import { Ability } from "../../../../types/enum/Ability"
-import { Pkm } from "../../../../types/enum/Pokemon"
-import GameScene from "../scenes/game-scene"
-import { addWanderingPokemon } from "./pokemon"
-import { DEPTH } from "../depths"
-
-const SHARDS_PER_ENCOUNTER = 50
-
-export default class UnownManager {
-  scene: GameScene
-
-  constructor(scene: GameScene) {
-    this.scene = scene
-  }
-
-  addWanderingUnown(pkm: Pkm, id: string) {
-    addWanderingPokemon(this.scene, id, pkm, (unown, id, pointer, tween) => {
-      this.scene.room?.send(Transfer.UNOWN_WANDERING, { id })
-      this.displayShardGain([pointer.x, pointer.y], unown.index)
-      unown.destroy()
-      tween.destroy()
-    })
-  }
-
-  displayShardGain(coordinates: number[], index: string) {
-    const textStyle = {
-      fontSize: "25px",
-      fontFamily: "Verdana",
-      color: "#fff",
-      align: "center",
-      strokeThickness: 2,
-      stroke: "#000"
-    }
-
-    const image = this.scene.add.existing(
-      new GameObjects.Image(this.scene, 0, 0, `portrait-${index}`)
-        .setScale(0.5, 0.5)
-        .setOrigin(0, 0)
-    )
-    const text = this.scene.add.existing(
-      new GameObjects.Text(
-        this.scene,
-        25,
-        0,
-        SHARDS_PER_ENCOUNTER.toString(),
-        textStyle
-      )
-    )
-    image.setDepth(DEPTH.TEXT_MINOR)
-    text.setDepth(DEPTH.TEXT)
-
-    const container = this.scene.add.existing(
-      new GameObjects.Container(
-        this.scene,
-        coordinates[0],
-        coordinates[1] - 50,
-        [text, image]
-      )
-    )
-
-    this.scene.add.tween({
-      targets: [container],
-      ease: "linear",
-      duration: 1500,
-      delay: 0,
-      alpha: {
-        getStart: () => 1,
-        getEnd: () => 0
-      },
-      y: {
-        getStart: () => coordinates[1] - 50,
-        getEnd: () => coordinates[1] - 110
-      },
-      onComplete: () => {
-        container.destroy()
-      }
-    })
-  }
-}
-
-export const UNOWNS_PER_ABILITY = new Map([
-  [
-    Ability.HIDDEN_POWER_A,
-    [Pkm.UNOWN_A, Pkm.UNOWN_B, Pkm.UNOWN_R, Pkm.UNOWN_A]
-  ],
-  [
-    Ability.HIDDEN_POWER_B,
-    [Pkm.UNOWN_B, Pkm.UNOWN_U, Pkm.UNOWN_R, Pkm.UNOWN_N]
-  ],
-  [
-    Ability.HIDDEN_POWER_C,
-    [Pkm.UNOWN_C, Pkm.UNOWN_O, Pkm.UNOWN_I, Pkm.UNOWN_N]
-  ],
-  [
-    Ability.HIDDEN_POWER_D,
-    [Pkm.UNOWN_D, Pkm.UNOWN_I, Pkm.UNOWN_T, Pkm.UNOWN_O]
-  ],
-  [
-    Ability.HIDDEN_POWER_E,
-    [Pkm.UNOWN_E, Pkm.UNOWN_G, Pkm.UNOWN_G, Pkm.UNOWN_S]
-  ],
-  [
-    Ability.HIDDEN_POWER_F,
-    [Pkm.UNOWN_F, Pkm.UNOWN_I, Pkm.UNOWN_S, Pkm.UNOWN_H]
-  ],
-  [
-    Ability.HIDDEN_POWER_G,
-    [Pkm.UNOWN_G, Pkm.UNOWN_O, Pkm.UNOWN_L, Pkm.UNOWN_D]
-  ],
-  [
-    Ability.HIDDEN_POWER_H,
-    [Pkm.UNOWN_H, Pkm.UNOWN_E, Pkm.UNOWN_A, Pkm.UNOWN_L]
-  ],
-  [
-    Ability.HIDDEN_POWER_I,
-    [Pkm.UNOWN_I, Pkm.UNOWN_T, Pkm.UNOWN_E, Pkm.UNOWN_M]
-  ],
-  [
-    Ability.HIDDEN_POWER_J,
-    [Pkm.UNOWN_J, Pkm.UNOWN_A, Pkm.UNOWN_W, Pkm.UNOWN_S]
-  ],
-  [
-    Ability.HIDDEN_POWER_K,
-    [Pkm.UNOWN_K, Pkm.UNOWN_I, Pkm.UNOWN_C, Pkm.UNOWN_K]
-  ],
-  [
-    Ability.HIDDEN_POWER_L,
-    [Pkm.UNOWN_L, Pkm.UNOWN_O, Pkm.UNOWN_C, Pkm.UNOWN_K]
-  ],
-  [
-    Ability.HIDDEN_POWER_M,
-    [Pkm.UNOWN_M, Pkm.UNOWN_A, Pkm.UNOWN_N, Pkm.UNOWN_A]
-  ],
-  [
-    Ability.HIDDEN_POWER_N,
-    [Pkm.UNOWN_N, Pkm.UNOWN_U, Pkm.UNOWN_K, Pkm.UNOWN_E]
-  ],
-  [
-    Ability.HIDDEN_POWER_O,
-    [Pkm.UNOWN_O, Pkm.UNOWN_V, Pkm.UNOWN_E, Pkm.UNOWN_N]
-  ],
-  [
-    Ability.HIDDEN_POWER_P,
-    [Pkm.UNOWN_P, Pkm.UNOWN_E, Pkm.UNOWN_S, Pkm.UNOWN_T]
-  ],
-  [
-    Ability.HIDDEN_POWER_Q,
-    [Pkm.UNOWN_Q, Pkm.UNOWN_U, Pkm.UNOWN_I, Pkm.UNOWN_T]
-  ],
-  [
-    Ability.HIDDEN_POWER_R,
-    [Pkm.UNOWN_R, Pkm.UNOWN_O, Pkm.UNOWN_L, Pkm.UNOWN_L]
-  ],
-  [
-    Ability.HIDDEN_POWER_S,
-    [Pkm.UNOWN_S, Pkm.UNOWN_T, Pkm.UNOWN_O, Pkm.UNOWN_P]
-  ],
-  [
-    Ability.HIDDEN_POWER_T,
-    [Pkm.UNOWN_T, Pkm.UNOWN_R, Pkm.UNOWN_E, Pkm.UNOWN_E]
-  ],
-  [
-    Ability.HIDDEN_POWER_U,
-    [Pkm.UNOWN_U, Pkm.UNOWN_X, Pkm.UNOWN_I, Pkm.UNOWN_E]
-  ],
-  [
-    Ability.HIDDEN_POWER_V,
-    [Pkm.UNOWN_V, Pkm.UNOWN_O, Pkm.UNOWN_L, Pkm.UNOWN_T]
-  ],
-  [
-    Ability.HIDDEN_POWER_W,
-    [Pkm.UNOWN_W, Pkm.UNOWN_I, Pkm.UNOWN_S, Pkm.UNOWN_H]
-  ],
-  [
-    Ability.HIDDEN_POWER_X,
-    [Pkm.UNOWN_X, Pkm.UNOWN_R, Pkm.UNOWN_A, Pkm.UNOWN_Y]
-  ],
-  [
-    Ability.HIDDEN_POWER_Y,
-    [Pkm.UNOWN_Y, Pkm.UNOWN_O, Pkm.UNOWN_G, Pkm.UNOWN_A]
-  ],
-  [
-    Ability.HIDDEN_POWER_Z,
-    [Pkm.UNOWN_Z, Pkm.UNOWN_Z, Pkm.UNOWN_Z, Pkm.UNOWN_Z]
-  ],
-  [
-    Ability.HIDDEN_POWER_EM,
-    [
-      Pkm.UNOWN_EXCLAMATION,
-      Pkm.UNOWN_EXCLAMATION,
-      Pkm.UNOWN_EXCLAMATION,
-      Pkm.UNOWN_EXCLAMATION
-    ]
-  ],
-  [
-    Ability.HIDDEN_POWER_QM,
-    [
-      Pkm.UNOWN_QUESTION,
-      Pkm.UNOWN_QUESTION,
-      Pkm.UNOWN_QUESTION,
-      Pkm.UNOWN_QUESTION
-    ]
-  ]
-])
